+++ src/components/Poptip/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withState } from 'recompose';
 import { waitForSelector } from '../../utils/poll';
+import classNames from 'classnames';
 
 import { Tooltip } from 'react-tippy';
 
@@ -38,6 +39,8 @@ export default class Poptip extends React.Component {
     hideDelay: PropTypes.number,
     /** should the poptip content be animated | default => true */
     isAnimated: PropTypes.bool,
+    /** should the poptip be displayed inline */
+    isInline: PropTypes.bool,
     /** position value options | default => top */
     position: PropTypes.oneOf([
       'top',
@@ -66,6 +69,7 @@ export default class Poptip extends React.Component {
     disable: false,
     hideDelay: 0,
     isAnimated: true,
+    isInline: false,
     position: 'top',
     theme: 'dark',
   };
@@ -78,10 +82,15 @@ export default class Poptip extends React.Component {
       disable,
       hideDelay,
       isAnimated,
+      isInline,
       position,
       trigger,
       theme,
     } = this.props;
+
+    const wrapperInline = classNames({
+      'display--inline': isInline,
+    });
 
     return (
       <Tooltip
@@ -93,7 +102,7 @@ export default class Poptip extends React.Component {
         position={ position }
         theme={ theme }
         trigger={ trigger }>
-        <div>{ children }</div>
+        <div className={ wrapperInline }>{ children }</div>
       </Tooltip>
     );
   };
@@ -103,4 +112,3 @@ export default class Poptip extends React.Component {
     return bodyDefined ? this.renderPoptip() : null;
   }
 }
-